% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/election_data.R
\name{election_data}
\alias{election_data}
\title{Connect to web API on Swedish election results}
\usage{
election_data(level, type)
}
\arguments{
\item{level}{characters}

\item{type}{characters}
}
\value{
put what the function returns here
}
\description{
This function allows the user to connect to the web API provided by Valmyndigheten to access the election results from 2014.
}
