\name{SwedishElectionData-package}
\alias{SwedishElectionData-package}
\alias{SwedishElectionData}
\docType{package}
\title{
\packageTitle{SwedishElectionData}
}
\description{
\packageDescription{SwedishElectionData}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SwedishElectionData}
\packageIndices{SwedishElectionData}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{SwedishElectionData}

Maintainer: \packageMaintainer{SwedishElectionData}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
