% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/election_data.R
\name{election_data}
\alias{election_data}
\title{Connect to web API on Swedish election results}
\usage{
election_data(level, type)
}
\arguments{
\item{...}{Numeric, complex, or logical vectors.}

\item{na.rm}{A logical scalar. Should missing values (including NaN)
be removed?}
}
\value{

}
\description{
This function allows the user to connect to the web API provided by Valmyndigheten to access the election results from 2014.
}
